challenge 5/requirements.txt
azure-quantum==0.18.2109.162713
numpy==1.21.2